Task dependencies are resolved relative to their parent
Previously task dependency search began at the dependent node itself. This was erroneous; consider the following task list, wherein a:b:c depends on x:y:

    a
    a:b:c => x:y
    a:b:c:x:y
    x:y

When task a:b:c is invoked, resolving the dependency on x:y relative to c will return a:b:c:x:y - this is not the correct outcome. Instead, the search must begin at a:b and continue up the tree until the root node is reached, at which point the correct x:y will be located.